refactor: seperate functions and files to fit norm standart
