+++ src/bonus/sprite/sort.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   sprite_sort.c                                      :+:      :+:    :+:   */
+/*   sort.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: hademirc <hademirc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/29 21:00:00 by hademirc          #+#    #+#             */
-/*   Updated: 2025/09/29 22:25:12 by hademirc         ###   ########.fr       */
+/*   Updated: 2025/10/01 18:47:28 by hademirc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
